+++ src/Concerns/Reporters.php
@@ -13,22 +13,6 @@ trait Reporters
         return EnumReporter::get();
     }
 
-    /**
-     * @deprecated
-     */
-    public static function makeOrReport(int|string|null $key, BackedEnum $context = null): ?self
-    {
-        return self::getOrReport($key, $context);
-    }
-
-    /**
-     * @deprecated
-     */
-    public static function makeOrReportArray(iterable $keys, BackedEnum $context = null): array
-    {
-        return self::getOrReportArray($keys, $context);
-    }
-
     public static function getOrReport(int|string|null $key, BackedEnum $context = null): ?self
     {
         return EnumReporter::getOrReport(self::class, $key, $context, self::reporter());