+++ src/Enums/LogLevel.php
@@ -6,6 +6,9 @@ use Henzeb\Enumhancer\Concerns\Configure;
 use Henzeb\Enumhancer\Concerns\Enhancers;
 use Henzeb\Enumhancer\Concerns\Macros;
 
+/**
+ * @uses Enhancers<self>
+ */
 enum LogLevel
 {
     use Enhancers, Configure, Macros;