+++ src/Helpers/Mappers/EnumArrayMapper.php
@@ -3,12 +3,17 @@
 namespace Henzeb\Enumhancer\Helpers\Mappers;
 
 use Henzeb\Enumhancer\Contracts\Mapper;
+use UnitEnum;
 
 final class EnumArrayMapper extends Mapper
 {
-    public function __construct(private array $mappable)
+    /**
+     * @param array<string|int,string|int|UnitEnum|array<string|int,string|int|UnitEnum>> $mappable
+     */
+    public function __construct(private readonly array $mappable)
     {
     }
+
 
     protected function mappable(): array
     {