+++ tests/Fixtures/IntBackedEnum.php
@@ -5,12 +5,10 @@ namespace Henzeb\Enumhancer\Tests\Fixtures;
 
 
 use Henzeb\Enumhancer\Concerns\Bitmasks;
-use Henzeb\Enumhancer\Concerns\Configure;
+use Henzeb\Enumhancer\Concerns\Comparison;
 use Henzeb\Enumhancer\Concerns\ConfigureDefaults;
 use Henzeb\Enumhancer\Concerns\Defaults;
 use Henzeb\Enumhancer\Concerns\Getters;
-use Henzeb\Enumhancer\Concerns\Makers;
-use Henzeb\Enumhancer\Concerns\Comparison;
 
 /**
  * @method isTest() bool
@@ -22,7 +20,7 @@ use Henzeb\Enumhancer\Concerns\Comparison;
  */
 enum IntBackedEnum: int
 {
-    use Makers, Getters, Comparison, Bitmasks, ConfigureDefaults, Defaults;
+    use Getters, Comparison, Bitmasks, ConfigureDefaults, Defaults;
 
     case TEST = 0;
     case TEST_2 = 1;