+++ tests/Fixtures/UnitEnums/Getters/GetUnitEnum.php
@@ -4,11 +4,10 @@ namespace Henzeb\Enumhancer\Tests\Fixtures\UnitEnums\Getters;
 
 use Henzeb\Enumhancer\Concerns\From;
 use Henzeb\Enumhancer\Concerns\Getters;
-use Henzeb\Enumhancer\Concerns\Makers;
 
 enum GetUnitEnum
 {
-    use Makers, Getters, From;
+    use Getters, From;
 
     case Zero;
     case One;